+++ RELEASE_NOTES.md
@@ -3,6 +3,9 @@
 ## 0.1.37
 __New feature__:
 - Export all tables in DDL2YML generation
+- Include table & column names when meeting unknown column type in JDBC source schema
+- Better logging on forced conversion in DDL2YML
+
 __Bug Fix__: